# Conflicts:
#	mosquitto/data/mosquitto.db

+++ Backend/src/main/kotlin/org/pin/backend/dto/PointDeltaDTO.kt
@@ -1,8 +1,8 @@
 package org.pin.backend.dto
 
 data class PointDeltaDTO(
-    val x: Short,
-    val y: Short,
+    val x: Float,
+    val y: Float,
     val size: Float,
     val color: Byte,
 )
